moving newFeature related code from serializePR
