+++ parsers/shared/src/test/scala/sigmastate/lang/LangTests.scala
@@ -79,4 +79,10 @@ trait LangTests extends Matchers with NegativeTesting {
         node
     }))(tree)
   }
+
+  /** Execute the given `block` having `version` as both activated and ErgoTree version. */
+  def runWithVersion[T](version: Byte)(block: => T): T = {
+    VersionContext.withVersions(version, version)(block)
+  }
+
 }
