+++ sc/shared/src/test/scala/sigmastate/lang/SigmaBinderTest.scala
@@ -10,10 +10,12 @@ import sigma.ast.syntax.SValue
 import sigmastate._
 import sigmastate.interpreter.Interpreter.ScriptEnv
 import SigmaPredef.PredefinedFuncRegistry
+import sigma.VersionContext
 import sigma.ast.syntax._
 import sigma.compiler.phases.SigmaBinder
 import sigma.eval.SigmaDsl
 import sigma.exceptions.BinderException
+import sigmastate.helpers.SigmaPPrint
 
 class SigmaBinderTest extends AnyPropSpec with ScalaCheckPropertyChecks with Matchers with LangTests {
   import StdSigmaBuilder._
@@ -27,6 +29,17 @@ class SigmaBinderTest extends AnyPropSpec with ScalaCheckPropertyChecks with Mat
     res.sourceContext.isDefined shouldBe true
     assertSrcCtxForAllNodes(res)
     res
+  }
+
+  /** Checks that parsing and binding results in the expected value.
+    * @return the inferred type of the expression
+    */
+  def checkBound(env: ScriptEnv, x: String, expected: SValue) = {
+    val bound = bind(env, x)
+    if (expected != bound) {
+      SigmaPPrint.pprintln(bound, width = 100)
+    }
+    bound shouldBe expected
   }
 
   private def fail(env: ScriptEnv, x: String, expectedLine: Int, expectedCol: Int): Unit = {
